opwen_email_client/requirements.txt
Babel==2.8.0
Flask-BabelEx==0.9.4
Flask-Caching==1.8.0
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-Testing==0.8.0
Flask-WTF==0.14.3
Flask==1.1.2
SQLAlchemy==1.3.16
WTForms==2.3.1
email-validator==1.0.5
Werkzeug==0.16.1  # pyup: ignore
apache-libcloud==3.0.0
bcrypt==3.1.7
beautifulsoup4==4.9.0
cached-property==1.5.1
celery[sqlalchemy]==4.4.2
environs==7.4.0
gunicorn==20.0.4
mkwvconf==0.1.1
passlib==1.7.2
python-crontab==2.4.2
requests==2.23.0
typing==3.7.4.1
tzlocal==2.0.0
watchdog==0.10.2
xtarfile[zstd]==0.0.2
Pillow==7.1.2
